+++ src/Kanvas/Users/Jobs/OnBoardingJob.php
@@ -13,13 +13,22 @@ use Illuminate\Foundation\Bus\Dispatchable;
 use Illuminate\Queue\InteractsWithQueue;
 use Illuminate\Queue\SerializesModels;
 use Illuminate\Support\Facades\Auth;
+use Kanvas\Apps\Models\Apps;
 use Kanvas\Companies\Models\Companies;
 use Kanvas\Companies\Models\CompaniesBranches;
 use Kanvas\Enums\AppSettingsEnums;
 use Kanvas\Event\Support\Enums\EventSetupTypeEnum;
 use Kanvas\Event\Support\Setup as EventSetup;
+use Kanvas\Exceptions\InternalServerErrorException;
 use Kanvas\Guild\Support\Setup as GuildSetup;
 use Kanvas\Inventory\Support\Setup as InventorySetup;
+use Kanvas\Workflow\Enums\IntegrationsEnum;
+use Kanvas\Workflow\Enums\StatusEnum;
+use Kanvas\Workflow\Integrations\Actions\CreateIntegrationCompanyAction;
+use Kanvas\Workflow\Integrations\DataTransferObject\IntegrationsCompany as IntegrationsCompanyDto;
+use Kanvas\Workflow\Integrations\Models\Status;
+use Kanvas\Workflow\Models\Integrations;
+use Throwable;
 
 class OnBoardingJob implements ShouldQueue
 {
@@ -35,7 +44,7 @@ class OnBoardingJob implements ShouldQueue
     public function __construct(
         public UserInterface $user,
         public CompaniesBranches $branch,
-        public AppInterface $app
+        public Apps $app
     ) {
     }
 
@@ -79,6 +88,12 @@ class OnBoardingJob implements ShouldQueue
                 $this->user,
                 $company
             ))->run();
+
+            try {
+                $this->setupDefaultIntegration($company, $this->app);
+            } catch (Throwable $e) {
+                report($e);
+            }
         }
 
         if ($runOnboardingEvent) {
@@ -93,5 +108,37 @@ class OnBoardingJob implements ShouldQueue
                 $eventSetupType
             )->run();
         }
+    }
+
+    private function setupDefaultIntegration(Companies $company, Apps $app): void
+    {
+        $integration = Integrations::getByName(IntegrationsEnum::INTERNAL->value);
+        $defaultRegion = $company->defaultRegion()->first();
+
+        if (! $integration || ! $defaultRegion) {
+            return;
+        }
+
+        $integrationDto = new IntegrationsCompanyDto(
+            integration: $integration,
+            region: $defaultRegion,
+            company: $company,
+            config: [],
+            app: $app
+        );
+
+        if (! class_exists($handler = $integration->handler)) {
+            throw new InternalServerErrorException('Handler Class not found.');
+        }
+
+        $status = Status::where('slug', StatusEnum::ACTIVE->value)
+                        ->where('apps_id', 0)
+                        ->first();
+
+        new CreateIntegrationCompanyAction(
+            $integrationDto,
+            $company->user,
+            $status
+        )->execute();
     }
 }
